Move various component protocols into seperate files
