+++ src/ikigai/typing/protocol/generic.py
@@ -1,0 +1,11 @@
+# SPDX-FileCopyrightText: 2024-present Harsh Parekh <harsh@ikigailabs.io>
+#
+# SPDX-License-Identifier: MIT
+
+from __future__ import annotations
+
+from typing import Protocol
+
+
+class Named(Protocol):
+    name: str
